# backend/docs/swagger.yaml
basePath: /api
definitions:
  controller.createTransactionReq:
    properties:
      amount:
        minimum: 0
        type: integer
      asset:
        type: string
      category:
        type: string
      date:
        type: string
      description:
        type: string
      flow:
        type: string
      notes:
        type: string
      quantity:
        type: number
      subtype:
        type: string
    required:
    - amount
    - category
    - date
    - description
    - flow
    type: object
  controller.updateTransactionReq:
    properties:
      amount:
        type: integer
      asset:
        type: string
      category:
        type: string
      date:
        type: string
      description:
        type: string
      flow:
        type: string
      notes:
        type: string
      quantity:
        type: number
      subtype:
        type: string
    type: object
  controller.upsertBudgetReq:
    properties:
      amount:
        minimum: 0
        type: integer
      category:
        type: string
      month:
        type: integer
      year:
        type: integer
    required:
    - category
    - year
    type: object
  domain.AnnualReport:
    properties:
      category_totals:
        items:
          $ref: '#/definitions/domain.CategorySummary'
        type: array
      kpis:
        $ref: '#/definitions/domain.KPIReport'
      monthly_trend:
        items:
          $ref: '#/definitions/domain.MonthlyReport'
        type: array
      year:
        type: integer
    type: object
  domain.Budget:
    properties:
      amount:
        type: integer
      category:
        type: string
      id:
        type: integer
      month:
        type: integer
      year:
        type: integer
    type: object
  domain.CategorySummary:
    properties:
      budget:
        type: integer
      category:
        type: string
      pct_used:
        type: number
      total:
        type: integer
      transactions:
        type: integer
    type: object
  domain.CreditCardItem:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      installment_current:
        type: integer
      installment_total:
        type: integer
      item_type:
        type: string
      statement_id:
        type: integer
    type: object
  domain.CreditCardStatement:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/domain.CreditCardItem'
        type: array
      period_from:
        type: string
      period_to:
        type: string
      total_amount:
        type: integer
    type: object
  domain.Insight:
    properties:
      category:
        type: string
      message:
        type: string
      type:
        type: string
    type: object
  domain.KPIReport:
    properties:
      cash_balance:
        type: integer
      cost_of_living:
        type: number
      expenses_ytd:
        type: integer
      fixed_expenses:
        type: integer
      income_ytd:
        type: integer
      investment_rate:
        type: number
      investments_ytd:
        type: integer
      net_worth:
        type: integer
    type: object
  domain.MonthlyReport:
    properties:
      balance:
        type: integer
      by_category:
        items:
          $ref: '#/definitions/domain.CategorySummary'
        type: array
      by_subtype:
        additionalProperties:
          format: int64
          type: integer
        type: object
      expenses:
        type: integer
      income:
        type: integer
      investments:
        type: integer
      month:
        type: integer
      savings_rate:
        type: number
      top_expenses:
        items:
          $ref: '#/definitions/domain.Transaction'
        type: array
      year:
        type: integer
    type: object
  domain.Transaction:
    properties:
      amount:
        type: integer
      asset:
        type: string
      category:
        type: string
      cc_statement_id:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      date:
        type: string
      description:
        type: string
      flow:
        type: string
      id:
        type: integer
      notes:
        type: string
      quantity:
        type: number
      source:
        type: string
      subtype:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Personal finance API — transactions, budgets, reports, and financial
    insights.
  title: Arasaka Finance API
  version: "1.0"
paths:
  /budgets:
    get:
      parameters:
      - description: Filter by year
        in: query
        name: year
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Budget'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List budgets
      tags:
      - budgets
    put:
      consumes:
      - application/json
      parameters:
      - description: Budget payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.upsertBudgetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Budget'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create or update a budget
      tags:
      - budgets
  /credit-card/statements:
    get:
      description: Returns all billing statements across national and international
        accounts, ordered by most recent period first.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.CreditCardStatement'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List credit card statements
      tags:
      - credit-card
  /credit-card/statements/{id}:
    get:
      description: Returns a single billing statement including all purchases, installments,
        commissions, and payments.
      parameters:
      - description: Statement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CreditCardStatement'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a credit card statement with its line items
      tags:
      - credit-card
  /insights:
    get:
      description: Returns rule-based alerts comparing current month data against
        historical averages and budgets.
      parameters:
      - description: Year (defaults to current)
        in: query
        name: year
        type: integer
      - description: Month 1-12 (defaults to current)
        in: query
        name: month
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Insight'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Financial insights for a month
      tags:
      - insights
  /reports/annual:
    get:
      description: Returns KPIs, monthly income/expense trend, and category totals
        for the given year
      parameters:
      - description: Year (defaults to current)
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AnnualReport'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Full-year financial overview
      tags:
      - reports
  /reports/budget-vs-actual:
    get:
      parameters:
      - description: Year (defaults to current)
        in: query
        name: year
        type: integer
      - description: Month 1-12 (defaults to current)
        in: query
        name: month
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.CategorySummary'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Budget vs actual spending
      tags:
      - reports
  /reports/kpis:
    get:
      parameters:
      - description: Year (defaults to current)
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.KPIReport'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Annual KPIs
      tags:
      - reports
  /reports/monthly:
    get:
      parameters:
      - description: Year (defaults to current)
        in: query
        name: year
        type: integer
      - description: Month 1-12 (defaults to current)
        in: query
        name: month
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MonthlyReport'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Monthly financial report
      tags:
      - reports
  /reports/trend:
    get:
      parameters:
      - description: Year (defaults to current)
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.MonthlyReport'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Monthly spending/income trend for a year
      tags:
      - reports
  /transactions:
    get:
      parameters:
      - description: Filter by year (e.g. 2026)
        in: query
        name: year
        type: string
      - description: Filter by month (1-12)
        in: query
        name: month
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by flow (income|expense|investment)
        in: query
        name: flow
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Transaction'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      parameters:
      - description: Transaction payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.createTransactionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Transaction'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a transaction
      tags:
      - transactions
  /transactions/{id}:
    delete:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a transaction
      tags:
      - transactions
    put:
      consumes:
      - application/json
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update (all optional)
        in: body
        name: body
        schema:
          $ref: '#/definitions/controller.updateTransactionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Transaction'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a transaction
      tags:
      - transactions
swagger: "2.0"